+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.2.3](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.2.2...@stream-io/video-react-native-sdk-0.2.3) (2023-11-01)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.4.2`
+* `@stream-io/video-react-bindings` updated to version `0.3.2`
 ### [0.2.2](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.2.1...@stream-io/video-react-native-sdk-0.2.2) (2023-10-30)
 
 ### Dependency Updates